refactored the horrendous win_printf function
win_printf used to be switch statements nested inside switch statements
nested inside a loop.   this was absolutely unmaintainable and was full
of silly bugs that were almsot impossible to see.  refactoring uncovered
them and made them jump right out at and smack you in the face.

also...

added an xabort() function for those special occasions
modified .clang-format which results in a few files with only format changes

changed ui_json.c help function to use my not patented smush compression
algorithm on the help data.  original in clear data is still there but
commented out.  smushing the data saved 300 bytes!  err well it ended up
only saving 64 bytes somehow but the data is 300 bytes smaller :)

+++ src/screen.c
@@ -220,8 +220,7 @@ void scr_add_backdrop(screen_t *scr)
 
         if(win != NULL)
         {
-            win->bdr_attrs.bytes[ATTR] =
-                (int8_t)(FG_GRAY | BG_GRAY | BOLD);
+            win->bdr_attrs.bytes[ATTR] = (FG_GRAY | BG_GRAY | BOLD);
             win->bdr_attrs.bytes[FG] = 13;
             win->bdr_attrs.bytes[BG] = 0;
             win->bdr_type = BDR_SINGLE;
@@ -271,7 +270,6 @@ static void scr_emit(screen_t *scr, int16_t index)
     p2 = &scr->buffer2[index];
 
     // are we about to write a wide character here..
-
 
     // mental note to self.  If you are about to read p1[1] make sure
     // p1 is not pointing to the last element of the array
